Refactor ElementAtOrAbsent tests
