+++ src/core-taggeds-optional/Optional.Tests/OptionalLinqExtensions.Tests/ElementAtOrAbsent.Enumerable.Int32.cs
@@ -1,0 +1,53 @@
+﻿using NUnit.Framework;
+using PrimeFuncPack.UnitTest;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using static PrimeFuncPack.UnitTest.TestData;
+
+namespace PrimeFuncPack.Core.Tests;
+
+partial class OptionalLinqExtensionsTests
+{
+    [Test]
+    public void ElementAtOrAbsentByInt_CollectionSourceIsNull_ExpectArgumentNullException()
+    {
+        IEnumerable<StructType> source = null!;
+        int index = 1;
+
+        var ex = Assert.Throws<ArgumentNullException>(() => _ = source.ElementAtOrAbsent(index));
+        Assert.AreEqual("source", ex?.ParamName);
+    }
+
+    [Test]
+    [TestCase(0)]
+    [TestCase(1)]
+    [TestCase(2)]
+    public void ElementAtOrAbsentByInt_CollectionIndexIsInRange_ExpectPresentItem(
+        int index)
+    {
+        var source = CreateCollection(PlusFifteenIdRefType, null, ZeroIdRefType);
+        var actual = source.ElementAtOrAbsent(index);
+
+        var expectedValue = source.ElementAt(index);
+        var expected = Optional<RefType?>.Present(expectedValue);
+
+        Assert.AreEqual(expected, actual);
+    }
+
+    [Test]
+    [TestCase(int.MinValue)]
+    [TestCase(-1)]
+    [TestCase(3)]
+    [TestCase(int.MaxValue)]
+    public void ElementAtOrAbsentByInt_CollectionIndexIsNotInRange_ExpectAbsent(
+        int index)
+    {
+        var source = CreateCollection(NullTextStructType, SomeTextStructType, default);
+
+        var actual = source.ElementAtOrAbsent(index);
+        var expected = Optional<StructType>.Absent;
+
+        Assert.AreEqual(expected, actual);
+    }
+}
