+++ src/core-taggeds-optional/Optional.Tests/OptionalLinqExtensions.Tests/ElementAtOrAbsent.Enumerable.Int64.cs
@@ -10,48 +10,6 @@ namespace PrimeFuncPack.Core.Tests;
 partial class OptionalLinqExtensionsTests
 {
     [Test]
-    public void ElementAtOrAbsentByInt_CollectionSourceIsNull_ExpectArgumentNullException()
-    {
-        IEnumerable<StructType> source = null!;
-        int index = 1;
-
-        var ex = Assert.Throws<ArgumentNullException>(() => _ = source.ElementAtOrAbsent(index));
-        Assert.AreEqual("source", ex?.ParamName);
-    }
-
-    [Test]
-    [TestCase(0)]
-    [TestCase(1)]
-    [TestCase(2)]
-    public void ElementAtOrAbsentByInt_CollectionIndexIsInRange_ExpectPresentItem(
-        int index)
-    {
-        var source = CreateCollection(PlusFifteenIdRefType, null, ZeroIdRefType);
-        var actual = source.ElementAtOrAbsent(index);
-
-        var expectedValue = source.ElementAt(index);
-        var expected = Optional<RefType?>.Present(expectedValue);
-
-        Assert.AreEqual(expected, actual);
-    }
-
-    [Test]
-    [TestCase(int.MinValue)]
-    [TestCase(-1)]
-    [TestCase(3)]
-    [TestCase(int.MaxValue)]
-    public void ElementAtOrAbsentByInt_CollectionIndexIsNotInRange_ExpectAbsent(
-        int index)
-    {
-        var source = CreateCollection(NullTextStructType, SomeTextStructType, default);
-
-        var actual = source.ElementAtOrAbsent(index);
-        var expected = Optional<StructType>.Absent;
-
-        Assert.AreEqual(expected, actual);
-    }
-
-    [Test]
     public void ElementAtOrAbsentByLong_CollectionSourceIsNull_ExpectArgumentNullException()
     {
         IEnumerable<StructType> source = null!;
